requirements.txt
Django==2.2.13
pytz==2020.4
django-cors-headers==2.1.0
djangorestframework==3.9.4
requests==2.20.0
django-filter==2.0.0
urllib3==1.24.2